include zk_c changes from previous release
- The zookeeper C lib from official source fails to build on smartos for
  a number of reasons, such as redefining htonl which is already
  provided on Solaris. Looks like the previous zk_c in the older release
  was not the official one as it contained a bunch of fixes for building
  on solaris that i can't find. The next version of the C lib 3.4.7 does include fixes for
  solaris but it has yet to be released.

  Carry on using previous version until 3.4.7 is released proper.

+++ src/node-zookeeper/zk_c/src/zk_log.c
@@ -133,7 +133,7 @@ void log_message(ZooLogLevel curLevel,int line,const char* funcName,
 #endif
     if(pid==0)pid=getpid();
 #ifndef THREADED
-    fprintf(LOGSTREAM, "%s:%d:%s@%s@%d: %s\n", time_now(get_time_buffer()),pid,
+    fprintf(LOGSTREAM, "%s:%d:%s@%s@%d: %s\n", time_now(get_time_buffer()),(int)pid,
             dbgLevelStr[curLevel],funcName,line,message);
 #else
 #ifdef WIN32
@@ -141,7 +141,7 @@ void log_message(ZooLogLevel curLevel,int line,const char* funcName,
             (unsigned long int)(pthread_self().thread_id),
             dbgLevelStr[curLevel],funcName,line,message);      
 #else
-    fprintf(LOGSTREAM, "%s:%d(0x%lx):%s@%s@%d: %s\n", time_now(get_time_buffer()),pid,
+    fprintf(LOGSTREAM, "%s:%d(0x%lx):%s@%s@%d: %s\n", time_now(get_time_buffer()),(int)pid,
             (unsigned long int)pthread_self(),
             dbgLevelStr[curLevel],funcName,line,message);      
 #endif
